fix: add python module resolution root and remove loading of env variables
