+++ cognitive_architecture/config.py
@@ -29,7 +29,7 @@ class Config:
     db_path = Path(__file__).resolve().parent / "database/data"
 
     vectordb: str = os.getenv("VECTORDB", "weaviate")
-    db_type: str = os.getenv("DB_TYPE", "postgres")
+    db_type: str = os.getenv("DB_TYPE", "sqlite")
     db_name: str = os.getenv("DB_NAME", "cognee.db")
     db_host: str = os.getenv("DB_HOST", "localhost")
     db_port: str = os.getenv("DB_PORT", "5432")
